new: Work in progress integration for Sponge.
This doesn't work in the slightest because common still uses some bukkit code. I'll have to make my own Yaml Configuration system.

The code is also jank, and untested because I haven't been able to get it to run yet. I'm just commiting to save progress. If I had to guess, it'd spam me with like a million errors anyway.