new: Mostly working integration for Sponge.
Sponge has so many problems that made it extremely difficult to work with. I'm not using their built-in asset system because of how horrible it is.

Depending on the server version you use cancelling the chat event will or won't work.

There is no way people actually use Sponge as a platform... it feels horrible to work with and barely works.